Add API and tests for value converters
Part of issue #242

As always, fluent API names subject to change in API review.

This PR also makes a change to never pass nulls to value converters. This makes it significantly easier to write custom converters that compose well with built-in converters and can be used for both PKs and FKs where the nullability is different.